requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML and Embeddings
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.5.2

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
markdown==3.5.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP and Networking
httpx==0.25.2
aiofiles==23.2.1
websockets==12.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Client Applications
kivy==2.2.0
kivymd==1.1.1
plyer==2.1.0